refactor: перемещен файл utils.js
